done task and show message when complete something

+++ SE1802_PRN212_Group6/ViewModels/Admin/ProductManangementViewModel.cs
@@ -114,18 +114,24 @@ namespace SE1802_PRN212_Group6.ViewModels.Admin
                 Temp.Category = _unitOfWork.CategoryRepository.GetById(Temp.Category.Id);
                 _unitOfWork.ProductRepository.Add(Temp);
                 _unitOfWork.SaveChanges();
+                Dialog.ShowSuccess("Added successfully");
                 Clear(obj);
             }
         }
 
         public void Delete(object obj)
         {
-            var get = _unitOfWork.ProductRepository.GetById(Select.Id);
-            if (get != null)
+            if (Dialog.ShowConfirm($"Are you sure you want to delete this product? (Id: {Select.Id})"))
             {
-                _unitOfWork.ProductRepository.Remove(get);
-                _unitOfWork.SaveChanges();
-                Clear(obj);
+
+                var get = _unitOfWork.ProductRepository.GetById(Select.Id);
+                if (get != null)
+                {
+                    _unitOfWork.ProductRepository.Remove(get);
+                    _unitOfWork.SaveChanges();
+                    Dialog.ShowSuccess("Deleted successfully");
+                    Clear(obj);
+                }
             }
         }
 
@@ -142,29 +148,36 @@ namespace SE1802_PRN212_Group6.ViewModels.Admin
                 get.Image = ImageDialog != null
                     ? ImageUtil.UpdateImage(nameof(Product), Select.Image, ImageDialog)
                     : get.Image;
-
-                _unitOfWork.ProductRepository.Update(get);
-                _unitOfWork.SaveChanges();
-                Clear(obj);
+                if (get.TryValidate())
+                {
+                    _unitOfWork.ProductRepository.Update(get);
+                    _unitOfWork.SaveChanges();
+                    Dialog.ShowSuccess("Updated successfully");
+                    Clear(obj);
+                }
             }
         }
 
         public void Restore(object obj)
         {
-            var get = _unitOfWork.ProductRepository.GetById(Select.Id);
-            if (get != null)
+            if (Dialog.ShowConfirm($"Are you sure you want to restore this product? (Id: {Select.Id})"))
             {
-                _unitOfWork.ProductRepository.Restore(get);
-                _unitOfWork.SaveChanges();
-                Clear(obj);
+                var get = _unitOfWork.ProductRepository.GetById(Select.Id);
+                if (get != null)
+                {
+                    _unitOfWork.ProductRepository.Restore(get);
+                    _unitOfWork.SaveChanges();
+                    Dialog.ShowSuccess("Restored successfully");
+                    Clear(obj);
+                }
             }
         }
 
         public void Clear(object obj)
         {
-            if (obj is ListView listRoom)
+            if (obj is ListView listProduct)
             {
-                listRoom.UnselectAll();
+                listProduct.UnselectAll();
             }
             Products.Clear();
             Load();
